Update issue 256
Acrivity API create を追加しました。


git-svn-id: http://aipo.googlecode.com/svn/trunk@691 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ orm/src/main/java/com/aipo/orm/model/social/auto/_Activity.java
@@ -81,11 +81,11 @@ public abstract class _Activity extends CayenneDataObject {
         return (String)readProperty("portletParams");
     }
 
-    public void setPriority(Integer priority) {
+    public void setPriority(Float priority) {
         writeProperty("priority", priority);
     }
-    public Integer getPriority() {
-        return (Integer)readProperty("priority");
+    public Float getPriority() {
+        return (Float)readProperty("priority");
     }
 
     public void setTitle(String title) {
